data_cache/requirements.txt
asn1crypto==1.5.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
filelock==3.8.2
idna==3.4
numpy==1.24.1
oscrypto==1.3.0
pandas==1.5.3
pycparser==2.21
pycryptodomex==3.16.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
snowflake-connector-python==2.9.0
typing_extensions==4.4.0
urllib3==1.26.13
